+++ Professions/Framework/Events/Input/ButtonPressed/RascalButtonPressedEvent.cs
@@ -27,7 +27,7 @@ internal sealed class RascalButtonPressedEvent(EventManager? manager = null)
 
         var player = Game1.player;
         if (Game1.activeClickableMenu is not null || player.CurrentTool is not Slingshot slingshot ||
-            slingshot.AttachmentSlotsCount != 2)
+            slingshot.AttachmentSlotsCount != 2 || slingshot.attachments.Length != 2)
         {
             return;
         }