+++ Professions/Framework/Patchers/Combat/SlingshotAttachPatcher.cs
@@ -26,7 +26,7 @@ internal sealed class SlingshotAttachPatcher : HarmonyPatcher
     [HarmonyPrefix]
     private static bool SlingshotAttachPrefix(Tool __instance, ref SObject? __result, SObject? o)
     {
-        if (__instance is not Slingshot { AttachmentSlotsCount: 2 })
+        if (__instance is not Slingshot { AttachmentSlotsCount: 2, attachments.Length: 2 })
         {
             return true; // run original logic
         }